+++ NEWS.md
@@ -7,6 +7,7 @@
   * Ensure library calls use quotes
   * Check margin notes & captions finish with a full stop
   * Add the course package version to the final page
+  * Remove old files in final/ when make final
 
 # 0.5.0
   * Extract latex tags using xparse from latex